+++ src/ClientApp/src/app/admin/import-polling-stations/imported-polling-stations-table/imported-polling-stations-table.component.scss
