+++ src/ClientApp/src/app/admin/import-polling-stations/imported-polling-stations-table/imported-polling-stations-table.component.ts
@@ -3,15 +3,15 @@ import { MatSort } from '@angular/material/sort';
 import { MatTableDataSource } from '@angular/material/table';
 import { Store, select } from '@ngrx/store';
 import { MatPaginator } from '@angular/material/paginator';
-import { Observable, merge, Subscription } from 'rxjs';
+import { merge, Subscription } from 'rxjs';
 import { tap, map } from 'rxjs/operators';
-import { ImportedPollingStation, ImportJobDetails, ImportedPollingStationsFilter } from 'src/app/services/data.service';
 import { ApplicationState } from 'src/app/state/reducers';
-import { LoadImportedPollingStationsAction, DeleteImportedPollingStationAction, UpdatePagination, ResetFilter, UpdateFilter } from 'src/app/state/actions';
-import { getImportedPollingStations, getImportedPollingStationsTotal, getCurrentImportJobDetails, getCurrentImportedPollingStationsFilter as getImportedPollingStationsFilter } from 'src/app/state/selectors';
 import { get } from 'lodash';
 import { MatDialog } from '@angular/material';
-import { PollingStationEditorComponent } from 'src/app/imported-polling-station-editor/polling-station-editor.component';
+import { PollingStationEditorComponent } from '../../imported-polling-station-editor/polling-station-editor.component';
+import { ImportedPollingStation, ImportedPollingStationsFilter, ImportJobDetails } from '../../services/polling-stations.service';
+import { UpdatePagination, UpdateFilter, ResetFilter, DeleteImportedPollingStationAction, LoadImportedPollingStationsAction } from '../../state/admin-actions';
+import { getImportedPollingStations, getCurrentImportJobDetails, getImportedPollingStationsTotal, getCurrentImportedPollingStationsFilter } from '../../state/admin-selectors';
 
 @Component({
   selector: 'app-imported-polling-stations-table',
@@ -52,7 +52,7 @@ export class ImportedPollingStationsTableComponent implements OnInit, OnDestroy,
       this.currentJob = currentJob;
     });
 
-    this.store.pipe(select(getImportedPollingStationsFilter)).subscribe(filter => this.filter = filter);
+    this.store.pipe(select(getCurrentImportedPollingStationsFilter)).subscribe(filter => this.filter = filter);
     this.store.pipe(select(getImportedPollingStationsTotal)).subscribe(total => this.importedPollingStationsTotal = total);
 
   }
