+++ src/ClientApp/src/app/admin/imported-polling-station-editor/polling-station-editor.component.scss
