+++ src/ClientApp/src/app/state/reducers.ts
@@ -1,16 +1,7 @@
 import { AppActions, ActionTypes } from './actions';
-import { VotingGuide, StaticData, PollingStationGroup, ImportedPollingStation, PaginatedResponse, ImportJobDetails, PaginationDetails, ImportedPollingStationsFilter as ImportedPollingStationsFilter } from '../services/data.service';
-import { AuthActions, AuthActionTypes } from './auth';
+import { VotingGuide, StaticData, PollingStationGroup } from '../services/data.service';
 
 import { LocationDetails } from '../services/here-address.service';
-import { isEqual } from 'lodash';
-
-export interface ImportPollingStationsState {
-    importedPollingStationsPagination: PaginationDetails;
-    importedPollingStationsFilter: ImportedPollingStationsFilter;
-    importJobDetails: ImportJobDetails;
-    importedPollingStations: PaginatedResponse<ImportedPollingStation>;
-}
 
 export interface ApplicationState {
     languages: string[];
@@ -21,25 +12,7 @@ export interface ApplicationState {
     error: string;
     pollingStations: PollingStationGroup[];
     selectedAddressDetails: LocationDetails;
-    auth: {
-        token: string;
-    };
-    import: ImportPollingStationsState
 }
-
-const DEFAULT_IPS_FILTER: ImportedPollingStationsFilter = {
-    address: '',
-    county: '',
-    institution: '',
-    locality: '',
-    pollingStationNumber: '',
-    resolvedAddressStatus: ''
-};
-
-const DEFAULT_PAGINATION: PaginationDetails = {
-    pageNumber: 0,
-    pageSize: 5
-};
 
 const initialState: ApplicationState = {
     languages: [],
@@ -50,15 +23,8 @@ const initialState: ApplicationState = {
     selectedLanguage: 'Ro', // change to enum
     pollingStations: [],
     selectedAddressDetails: undefined,
-    auth: { token: '' },
-    import: {
-        importJobDetails: undefined,
-        importedPollingStations: undefined,
-        importedPollingStationsFilter: DEFAULT_IPS_FILTER,
-        importedPollingStationsPagination: DEFAULT_PAGINATION
-    }
 };
-export function appStateReducer(state: ApplicationState = initialState, action: AppActions | AuthActions): ApplicationState {
+export function appStateReducer(state: ApplicationState = initialState, action: AppActions): ApplicationState {
     switch (action.type) {
         case ActionTypes.LOAD_DATA_DONE:
             const languageData = action.payload.data.content.find(x => x.language === state.selectedLanguage);
@@ -87,38 +53,6 @@ export function appStateReducer(state: ApplicationState = initialState, action: 
                 selectedLanguage: action.payload
             };
 
-        case ActionTypes.CLEAR_ERROR:
-            return {
-                ...state,
-                error: '',
-            };
-
-        case ActionTypes.UPDATE_DATA_ERROR:
-            return {
-                ...state,
-                error: action.error.detail
-            };
-
-        case AuthActionTypes.LOGIN_SUCCEEDED:
-            const { token } = action.payload;
-            return {
-                ...state,
-                error: '',
-                auth: { token }
-            };
-        case AuthActionTypes.LOGIN_FAILED:
-            return {
-                ...state,
-                error: action.payload,
-                auth: { token: '' }
-            };
-        case AuthActionTypes.LOGOUT:
-            return {
-                ...state,
-                error: '',
-                auth: { token: '' }
-            };
-
         case ActionTypes.LOAD_LOCATIONS_DONE:
             return {
                 ...state,
@@ -126,71 +60,6 @@ export function appStateReducer(state: ApplicationState = initialState, action: 
                 selectedAddressDetails: action.userLocation
             };
 
-        case ActionTypes.LOAD_IMPORT_JOB_DETAILS_DONE:
-            return {
-                ...state,
-                import: {
-                    ...state.import,
-                    importJobDetails: action.importJobDetails
-                },
-            };
-
-        case ActionTypes.LOAD_IMPORT_JOB_DETAILS_ERROR:
-            return {
-                ...state,
-                import: {
-                    ...state.import,
-                    importJobDetails: undefined,
-                },
-            };
-
-        case ActionTypes.LOAD_IPS_DONE:
-            return {
-                ...state,
-                import: {
-                    ...state.import,
-                    importedPollingStations: action.payload,
-                },
-            };
-        case ActionTypes.LOAD_IPS_ERROR:
-            return {
-                ...state,
-                import: {
-                    ...state.import,
-                    importedPollingStations: undefined,
-                },
-            };
-
-        case ActionTypes.RESET_FILTER:
-            return {
-                ...state,
-                import: {
-                    ...state.import,
-                    importedPollingStationsFilter: DEFAULT_IPS_FILTER
-                }
-            }
-
-        case ActionTypes.UPDATE_FILTER:
-            if (isEqual(state.import.importedPollingStationsFilter, action.payload)) {
-                return state;
-            }
-
-            return {
-                ...state,
-                import: {
-                    ...state.import,
-                    importedPollingStationsFilter: action.payload
-                }
-            }
-
-        case ActionTypes.UPDATE_PAGINATION:
-            return {
-                ...state,
-                import: {
-                    ...state.import,
-                    importedPollingStationsPagination: action.payload
-                }
-            }
         default:
             return state;
     }
