+++ src/ClientApp/src/app/voters-guide/voters-guide.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit, OnDestroy, ElementRef, ViewChild } from '@angular/core';
+import { Component, OnInit, OnDestroy } from '@angular/core';
 import { ApplicationState } from '../state/reducers';
 import { Store, select } from '@ngrx/store';
 import { Subscription, Observable } from 'rxjs';
