+++ src/Services/IneffectiveSearchService.cs
@@ -3,59 +3,49 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using GeoCoordinatePortable;
-using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.DependencyInjection;
 using VotRomania.Models;
 using VotRomania.Stores;
-using VotRomania.Stores.Entities;
 
 namespace VotRomania.Services
 {
     public class IneffectiveSearchService : IPollingStationSearchService
     {
-        private readonly VotRomaniaContext _context;
+        private readonly IServiceScopeFactory _serviceScopeFactory;
 
         private List<(GeoCoordinate Coordinates, PollingStationModel[] PollingStations)> _pollingStations;
 
         // TODO: Change to KDTree
-        public IneffectiveSearchService(VotRomaniaContext context)
+        public IneffectiveSearchService(IServiceScopeFactory serviceScopeFactory)
         {
-            _context = context;
+            _serviceScopeFactory = serviceScopeFactory;
             PrefillPollingStationsCache().GetAwaiter().GetResult();
         }
 
         private async Task PrefillPollingStationsCache()
         {
-            var pollingStations = _context.PollingStations
-                .Select(pollingStation => new PollingStationModel()
-                {
-                    Id = pollingStation.Id,
-                    Address = pollingStation.Address,
-                    Longitude = pollingStation.Longitude,
-                    Latitude = pollingStation.Latitude,
-                    County = pollingStation.County,
-                    PollingStationNumber = pollingStation.PollingStationNumber,
-                    Locality = pollingStation.Locality,
-                    Institution = pollingStation.Institution,
-                    AssignedAddresses = pollingStation.PollingStationAddresses.Select(a => MapToAssignedAddresses(a))
-                })
-                .ToListAsync();
 
-            _pollingStations = pollingStations
-                .GroupBy(
-                    x => new
-                    {
-                        lat = x.Latitude,
-                        lng = x.Longitude
-                    },
+            using (var scope = _serviceScopeFactory.CreateScope())
+            {
+                var pollingStationsRepository = scope.ServiceProvider.GetService<IPollingStationsRepository>();
+                var pollingStationsResult = await pollingStationsRepository.GetPollingStationsAsync();
+
+                if (pollingStationsResult.IsFailure)
+                {
+                    throw new ApplicationException("could not load polling stations");
+                }
+
+                _pollingStations = pollingStationsResult
+                    .Value
+                    .Results
+                    .GroupBy(
+                        x => new { lat = x.Latitude, lng = x.Longitude },
                         p => p,
-                        (key, pollingStationsInfos) => (new GeoCoordinate(key.lat, key.lng), pollingStationsInfos.ToArray())
-                        )
+                        (key, pollingStationsInfos) =>
+                            (new GeoCoordinate(key.lat, key.lng), pollingStationsInfos.ToArray())
+                    )
                     .ToList();
-            x => new { lat = x.Latitude, lng = x.Longitude },
-                    p => p,
-                    (key, pollingStationsInfos) => (new GeoCoordinate(key.lat, key.lng), pollingStationsInfos.ToArray())
-                )
-                .ToList();
+            }
         }
 
         public Task<PollingStationsGroupModel[]> GetNearestPollingStationsAsync(double latitude, double longitude)
