+++ src/Services/Location/IAddressLocationSearchService.cs
@@ -5,6 +5,6 @@ namespace VotRomania.Services.Location
 {
     public interface IAddressLocationSearchService
     {
-        Task<LocationSearchResult> FindCoordinates(string fullAddress);
+        Task<LocationSearchResult> FindCoordinates(string county, string fullAddress);
     }
 }
