+++ src/Stores/IImportJobsRepository.cs
@@ -15,5 +15,6 @@ namespace VotRomania.Stores
         Task<Result<JobStatusModel>> GetImportJobStatus(Guid jobId);
         Task<Result> UpdateJobStatus(Guid jobId, JobStatus jobJobStatus);
         Task<Result<JobStatusModel>> GetCurrentImportJob();
+        Task<Result> RestartJob(Guid jobId);
     }
 }
