+++ src/appsettings.Development.json
@@ -10,10 +10,10 @@
         "VotRoDBConnection": "Server=localhost;Port=5432;Database=VotRomania;User Id=docker;Password=docker;"
     },
     "HereMaps": {
-        "Token": "HERE_TOKEN"
+        "Token": "Um0LhLV4phI2QpCYrBCwmWgvdjmH6NFvd709PhMqsQg"
     },
     "Authentication": {
-        "Secret": ""
+        "Secret": "7b#GnBNA07Tjl!*#@1xs$7xlo#dpVDxtgcM6Hnm&*ktnL*iVvJx03#2$D1ZK*vVMqCYSoHUHPtOLGA!u0%bf6T$Qd2mpyy0KpUt"
     },
     "UserSettings": {
         "Users": [
